Remove redundant system spec
